Commented making Intersections to prevent Exception
If DijkstraAlgorithm cooperates with intersection it has to be uncommented

+++ patients-transport-center/src/main/java/pl/group2/optimizer/impl/algorithms/dijkstra/DijkstraAlgorithm.java
@@ -45,7 +45,7 @@ public class DijkstraAlgorithm {
     private Graph makeGraph() {
         Graph graph = new Graph(hospitals.getMaxId() + 1);
         for (Path path : paths.getList()) {
-            graph.addEdge(path.getFrom().getId(), path.getTo().getId(), (int) path.getDistance());
+            graph.addEdge(path.getFrom().getId(), path.getTo().getId(), path.getDistance());
         }
         return graph;
     }
